Retheme the Bytes, looking pretty reasonable really
